+++ src/Makevars.win
@@ -1,1 +1,1 @@
-OBJECTS=rapi.o subprocess.o sub-windows.o utf8.o tests.o
+OBJECTS=rapi.o subprocess.o sub-windows.o tests.o
